move `update_client` method to submodule
